+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Optimize gas limits per PoCo function. This will enable validators
   to add more finalize and contributeAndFinalize transactions in a block. (#117)
 - Remove unnecessary blockchain calls to diminish pressure on Ethereum JSON-RPC API. (#124)
+- Remove boolean to yes/no conversion methods from `TeeUtils`. (#126)
 
 ### Breaking API changes